requirements.txt
pillow
pycairo
materialyoucolor
materialshapes
asynckivy
kivymd2
pytest